SD-12.0 guardrails/requirements.txt
crewai
crewai_tools
guardrails-ai
pydantic
